move enableScrolling to Popup
