+++ js/ui/popup.d.ts
@@ -191,12 +191,6 @@ export interface dxPopupOptions<TComponent> extends dxOverlayOptions<TComponent>
     position?: PositionAlignment | PositionConfig | Function;
     /**
      * @docid
-     * @default false
-     * @public
-     */
-    preventScrollEvents?: boolean;
-    /**
-     * @docid
      * @default true
      * @public
      */
