+++ js/ui/validation_message.d.ts
@@ -17,13 +17,6 @@ export interface dxValidationMessageOptions extends dxOverlayOptions<dxValidatio
     boundary?: String | UserDefinedElement;
 
     offset?: object;
-
-    /**
-     * @docid
-     * @default false
-     * @public
-     */
-    preventScrollEvents?: boolean;
 }
 /** @namespace DevExpress.ui */
 export default class dxValidationMessage extends dxOverlay<dxValidationMessageOptions> { }
